Create ManyToManyRelationship.py
When we have a Many-to-Many relationship then there is a problem to define which column to be treated as Foreign Key. Hence to solve this problem, we have to create a Connector Table/ Junction Table.  This table contains those two columns which are not having a many-to-many relationship. These two columns will be Foreign Keys in the Connector Table.

One, thing to note here is that even though the records in the individual Foreign Key column gets repeated but the combination of those FKeys in any record will always be Unique.